new branch for connected messages.
The old code is in ab-old.  The new is in ab.  This is the start of
a major rewrite.  Tasklets are now used for all back-end work including
session IO and polling.

Current code compiles and can open a session and a connection to a
Logix-class PLC.

Checking in for safety.